Client prep

+++ src/Conversation.java
@@ -50,6 +50,14 @@ public class Conversation {
         this.addToFile = false;
         this.database = database;
 
+        for (Account account : participants) {
+            try {
+                this.database.getAccountByUsername(account.getUsername()).addToConversation(this);
+            } catch (AccountNotExistException e) {
+                e.printStackTrace();
+            }
+        }
+
         this.database.addToDatabase(this);
     }
 
@@ -125,7 +133,9 @@ public class Conversation {
         Account account = this.database.getAccountByUsername(username);
         participants.remove(account);
         account.removeConversation(this);
-        this.database.removeParticipantFromConversationFile(this.getConversationId(), username);
+        if (this.database.isServer()) {
+            this.database.removeParticipantFromConversationFile(this.getConversationId(), username);
+        }
     }
 
     public void addMessage(Message message) {
